requirements.txt
blessed==1.14.2
eventlet==0.20.0
flask==0.12.1
flask-cors==3.0.2
flask-socketio==2.8.6
requests==2.14.2
pyyaml==3.12
jinja2==2.9.6
pytest==3.1.2
pexpect==4.2.1
python-dotenv==0.6.4
jsonschema==2.6.0
click==6.7
